DriverProfile/requirements.txt
astral==2.2
atomicwrites==1.4.1
attrs==21.4.0
beautifulsoup4==4.11.1
bidict==0.22.0
biosppy==0.8.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coloredlogs==15.0.1
cycler==0.11.0
decorator==5.1.1
flatbuffers==2.0
fonttools==4.34.4
future==0.18.2
geographiclib==1.52
geojson==2.5.0
geopy==2.2.0
h5py==3.7.0
humanfriendly==10.0
idna==3.3
imbalanced-learn==0.9.1
importlib-metadata==4.12.0
iniconfig==1.1.1
install==1.3.5
joblib==1.1.0
jplephem==2.17
kiwisolver==1.4.4
llvmlite==0.38.1
lxml==4.9.1
matplotlib==3.5.2
mpmath==1.2.1
numpy==1.22.4
onnx==1.12.0
onnxconverter-common==1.9.0
onnxmltools==1.11.1
onnxruntime==1.12.0
opencv-python==4.6.0.66
overpy==0.6
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
plotly==5.9.0
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pyclustering==0.10.1.2
pyparsing==3.0.9
pyreadline3==3.4.1
pytest==7.1.2
pytest-sugar==0.9.5
python-dateutil==2.8.2
pytz==2022.1
ratelim==0.1.6
requests==2.28.1
scikit-learn==1.1.1
scikit-learn-extra==0.2.0
scipy==1.8.1
seaborn==0.11.2
sgp4==2.21
shortuuid==1.0.9
six==1.16.0
skl2onnx==1.11.2
sklearn-pandas==2.2.0
sklearn2pmml==0.85.0
soupsieve==2.3.2.post1
suntime==1.2.5
sympy==1.10.1
tenacity==8.0.1
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
xarray==2022.6.0
xgboost==1.6.1
zipp==3.8.1
